made a route for creating admins, updating admin status

+++ src/client/components/admin/UsersList.jsx
@@ -1,7 +1,13 @@
 import { useState, useEffect } from "react"
+import { motion } from "framer-motion"
 
 import {
-  Container, SimpleGrid, Button, Box, Text,
+  Container,
+  SimpleGrid,
+  Button,
+  Box,
+  Text,
+  Card,
 } from "@chakra-ui/react"
 import adminServices from "../../services/admin"
 
@@ -13,28 +19,41 @@ const UsersList = () => {
   }, [])
 
   const onRemove = async (id) => {
+    if (!window.confirm("Are you sure?")) return
     await adminServices.deleteUser(id)
     setUsers(users.filter((user) => user.id !== id))
+  }
+
+  const makeAdmin = async (id) => {
+    if (!window.confirm("Are you sure?")) return
+    await adminServices.makeAdmin(id)
+    setUsers(
+      users.map((user) => (user.id === id ? { ...user, isAdmin: true } : user))
+    )
   }
 
   return (
     <Container maxW="container.lg">
       <SimpleGrid columns={[1, 2, 3]} gap={6}>
-        {users.map((user) => (user.isAdmin ? null : (
-            <Box
-              boxShadow="md"
-              p="6"
-              textAlign="center"
-              rounded="md"
-              style={{ fontWeight: "bold" }}
-              key={user.id}
-            >
-              <Text fontSize="lg">{user.username}</Text>
-              <Button w="full" onClick={() => onRemove(user.id)}>
-                Remove
-              </Button>
-            </Box>
-        )))}
+        {users.map((user) => (
+          <Card key={user.id}>
+            <Text fontSize="lg">{user.username}</Text>
+
+            {user.isAdmin ? (
+              "Admin"
+            ) : (
+              <>
+                {" "}
+                <Button w="full" onClick={() => onRemove(user.id)}>
+                  Remove
+                </Button>
+                <Button w="full" onClick={() => makeAdmin(user.id)}>
+                  Make Admin
+                </Button>
+              </>
+            )}
+          </Card>
+        ))}
       </SimpleGrid>
     </Container>
   )
